use server error class
