+++ code/lib/core-events/package.json
@@ -81,6 +81,7 @@
     "ts-dedent": "^2.0.0"
   },
   "devDependencies": {
+    "chalk": "^4.1.0",
     "typescript": "^5.3.2"
   },
   "publishConfig": {
